Introduction of muting and flashbacks
The mute/unmute button is only present in the interrogation scene, not any of the flashbacks. The flashbacks are called, but do not return back to interrogation scene automatically after the video finishes playing. In addition, when they do return back (if the skip button is pressed (currently hidden)), the story resumes to after the call flashback.